mutator: expose function instead of tuple + move mutators to dedicated file
Instead of an iterable of mutator, expose a function that choose
randomly in a set of mutators.
This way, mutator is no longer different of other function module like
reproduction or selection, and will therefore be handled easily in the
rest of the code.
Also, it's not the client code that have to choose a mutator to apply on
units: mutators handle their responsibility now.