Found and fix a major misconception about data returned by stepping functions
Stepping functions does NOT return a tuple (pop, scored_pop) where
scored_pop is the pop mapped with scores.
Instead, scored_pop is the map of INDIVIDUALS OF THE PREVIOUS STEP
THAT ARE IN THE CURRENT STEP with their score.
This is a big difference that explains a non-working mmh until
this commit, that simply fixes the problem by putting pop instead of
scored_pop as output.

This also explains why lower score in populations are always so high :
because only the kept individuals are kept.

The byproduct of this commit is that scoring related data are now
inacessible from end-user.
Allowing a call-back in stepping function with all data in parameter
could be a good way to keep the current arch without sacrifying
too much data handling simplicity.